update test framework

+++ test/crosscheck.h
@@ -120,52 +120,52 @@
   * __CC_VALUES_ASSIGN handles copying the values under test into the result
   * struct for error output.
   */
- #define __CC_VALUES_ASSIGN(actual, expected, test_type)            \
-     if ((test_type) == test_type_char) {                           \
-         ccrt.exp = (test_values_t) {.char_val = (char)(expected)}; \
-         ccrt.act = (test_values_t) {.char_val = (char)(actual)};   \
-     } else if ((test_type) == test_type_float) {                   \
-         ccrt.exp = (test_values_t) {.float_val = (expected)};      \
-         ccrt.act = (test_values_t) {.float_val = (actual)};        \
-     } else if ((test_type) == test_type_double) {                  \
-         ccrt.exp = (test_values_t) {.double_val = (expected)};     \
-         ccrt.act = (test_values_t) {.double_val = (actual)};       \
-     } else if ((test_type) == test_type_long) {                    \
-         ccrt.exp = (test_values_t) {.long_val = (expected)};       \
-         ccrt.act = (test_values_t) {.long_val = (actual)};         \
-     } else if ((test_type) == test_type_long_long) {               \
-         ccrt.exp = (test_values_t) {.long_long_val = (expected)};  \
-         ccrt.act = (test_values_t) {.long_long_val = (actual)};    \
-     } else if ((test_type) == test_type_int) {                     \
-         ccrt.exp = (test_values_t) {.int_val = (expected)};        \
-         ccrt.act = (test_values_t) {.int_val = (actual)};          \
-     } else if ((test_type) == test_type_int8) {                    \
-         ccrt.exp = (test_values_t) {.int8_val = (expected)};       \
-         ccrt.act = (test_values_t) {.int8_val = (actual)};         \
-     } else if ((test_type) == test_type_int16) {                   \
-         ccrt.exp = (test_values_t) {.int16_val = (expected)};      \
-         ccrt.act = (test_values_t) {.int16_val = (actual)};        \
-     } else if ((test_type) == test_type_int32) {                   \
-         ccrt.exp = (test_values_t) {.int32_val = (expected)};      \
-         ccrt.act = (test_values_t) {.int32_val = (actual)};        \
-     } else if ((test_type) == test_type_int64) {                   \
-         ccrt.exp = (test_values_t) {.int64_val = (expected)};      \
-         ccrt.act = (test_values_t) {.int64_val = (actual)};        \
-     } else if ((test_type) == test_type_uint) {                    \
-         ccrt.exp = (test_values_t) {.uint_val = (expected)};       \
-         ccrt.act = (test_values_t) {.uint_val = (actual)};         \
-     } else if ((test_type) == test_type_uint8) {                   \
-         ccrt.exp = (test_values_t) {.uint8_val = (expected)};      \
-         ccrt.act = (test_values_t) {.uint8_val = (actual)};        \
-     } else if ((test_type) == test_type_uint16) {                  \
-         ccrt.exp = (test_values_t) {.uint16_val = (expected)};     \
-         ccrt.act = (test_values_t) {.uint16_val = (actual)};       \
-     } else if ((test_type) == test_type_uint32) {                  \
-         ccrt.exp = (test_values_t) {.uint32_val = (expected)};     \
-         ccrt.act = (test_values_t) {.uint32_val = (actual)};       \
-     } else if ((test_type) == test_type_uint64) {                  \
-         ccrt.exp = (test_values_t) {.uint64_val = (expected)};     \
-         ccrt.act = (test_values_t) {.uint64_val = (actual)};       \
+ #define __CC_VALUES_ASSIGN(actual, expected, test_type)                      \
+     if ((test_type) == test_type_char) {                                     \
+         ccrt.exp = (test_values_t) {.char_val = (char)(expected)};           \
+         ccrt.act = (test_values_t) {.char_val = (char)(actual)};             \
+     } else if ((test_type) == test_type_float) {                             \
+         ccrt.exp = (test_values_t) {.float_val = (float)(expected)};         \
+         ccrt.act = (test_values_t) {.float_val = (float)(actual)};           \
+     } else if ((test_type) == test_type_double) {                            \
+         ccrt.exp = (test_values_t) {.double_val = (double)(expected)};       \
+         ccrt.act = (test_values_t) {.double_val = (double)(actual)};         \
+     } else if ((test_type) == test_type_long) {                              \
+         ccrt.exp = (test_values_t) {.long_val = (long)(expected)};           \
+         ccrt.act = (test_values_t) {.long_val = (long)(actual)};             \
+     } else if ((test_type) == test_type_long_long) {                         \
+         ccrt.exp = (test_values_t) {.long_long_val = (long long)(expected)}; \
+         ccrt.act = (test_values_t) {.long_long_val = (long long)(actual)};   \
+     } else if ((test_type) == test_type_int) {                               \
+         ccrt.exp = (test_values_t) {.int_val = (int)(expected)};             \
+         ccrt.act = (test_values_t) {.int_val = (int)(actual)};               \
+     } else if ((test_type) == test_type_int8) {                              \
+         ccrt.exp = (test_values_t) {.int8_val = (int8_t)(expected)};         \
+         ccrt.act = (test_values_t) {.int8_val = (int8_t)(actual)};           \
+     } else if ((test_type) == test_type_int16) {                             \
+         ccrt.exp = (test_values_t) {.int16_val = (int16_t)(expected)};       \
+         ccrt.act = (test_values_t) {.int16_val = (int16_t)(actual)};         \
+     } else if ((test_type) == test_type_int32) {                             \
+         ccrt.exp = (test_values_t) {.int32_val = (int32_t)(expected)};       \
+         ccrt.act = (test_values_t) {.int32_val = (int32_t)(actual)};         \
+     } else if ((test_type) == test_type_int64) {                             \
+         ccrt.exp = (test_values_t) {.int64_val = (int64_t)(expected)};       \
+         ccrt.act = (test_values_t) {.int64_val = (int64_t)(actual)};         \
+     } else if ((test_type) == test_type_uint) {                              \
+         ccrt.exp = (test_values_t) {.uint_val = (unsigned int)(expected)};   \
+         ccrt.act = (test_values_t) {.uint_val = (unsigned int)(actual)};     \
+     } else if ((test_type) == test_type_uint8) {                             \
+         ccrt.exp = (test_values_t) {.uint8_val = (uint8_t)(expected)};       \
+         ccrt.act = (test_values_t) {.uint8_val = (uint8_t)(actual)};         \
+     } else if ((test_type) == test_type_uint16) {                            \
+         ccrt.exp = (test_values_t) {.uint16_val = (uint16_t)(expected)};     \
+         ccrt.act = (test_values_t) {.uint16_val = (uint16_t)(actual)};       \
+     } else if ((test_type) == test_type_uint32) {                            \
+         ccrt.exp = (test_values_t) {.uint32_val = (uint32_t)(expected)};     \
+         ccrt.act = (test_values_t) {.uint32_val = (uint32_t)(actual)};       \
+     } else if ((test_type) == test_type_uint64) {                            \
+         ccrt.exp = (test_values_t) {.uint64_val = (uint64_t)(expected)};     \
+         ccrt.act = (test_values_t) {.uint64_val = (uint64_t)(actual)};       \
      }
  
  /**
@@ -207,67 +207,67 @@
      } while (0)
  
  /* Type specific wrapper macros for equality */
- #define CC_ASSERT_CHAR_EQUAL(actual, expected)                            \
+ #define CC_ASSERT_CHAR_EQUAL(actual, expected)                               \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_char)
- #define CC_ASSERT_FLOAT_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_FLOAT_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_float)
- #define CC_ASSERT_DOUBLE_EQUAL(actual, expected)                          \
+ #define CC_ASSERT_DOUBLE_EQUAL(actual, expected)                             \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_double)
- #define CC_ASSERT_LONG_EQUAL(actual, expected)                            \
+ #define CC_ASSERT_LONG_EQUAL(actual, expected)                               \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_long)
- #define CC_ASSERT_LONG_LONG_EQUAL(actual, expected)                       \
+ #define CC_ASSERT_LONG_LONG_EQUAL(actual, expected)                          \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_long_long)
- #define CC_ASSERT_INT_EQUAL(actual, expected)                             \
+ #define CC_ASSERT_INT_EQUAL(actual, expected)                                \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_int)
- #define CC_ASSERT_INT8_EQUAL(actual, expected)                            \
+ #define CC_ASSERT_INT8_EQUAL(actual, expected)                               \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_int8)
- #define CC_ASSERT_INT16_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_INT16_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_int16)
- #define CC_ASSERT_INT32_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_INT32_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_int32)
- #define CC_ASSERT_INT64_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_INT64_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_int64)
- #define CC_ASSERT_UINT_EQUAL(actual, expected)                            \
+ #define CC_ASSERT_UINT_EQUAL(actual, expected)                               \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_uint)
- #define CC_ASSERT_UINT8_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_UINT8_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_uint8)
- #define CC_ASSERT_UINT16_EQUAL(actual, expected)                          \
+ #define CC_ASSERT_UINT16_EQUAL(actual, expected)                             \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_uint16)
- #define CC_ASSERT_UINT32_EQUAL(actual, expected)                          \
+ #define CC_ASSERT_UINT32_EQUAL(actual, expected)                             \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_uint32)
- #define CC_ASSERT_UINT64_EQUAL(actual, expected)                          \
+ #define CC_ASSERT_UINT64_EQUAL(actual, expected)                             \
      __CC_ASSERT_NUMBER_EQUAL_TYPE((actual), (expected), test_type_uint64)
  
  /* Type specific wrapper macros for in-equality */
- #define CC_ASSERT_CHAR_NOT_EQUAL(actual, expected)                            \
+ #define CC_ASSERT_CHAR_NOT_EQUAL(actual, expected)                               \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_char)
- #define CC_ASSERT_FLOAT_NOT_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_FLOAT_NOT_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_float)
- #define CC_ASSERT_DOUBLE_NOT_EQUAL(actual, expected)                          \
+ #define CC_ASSERT_DOUBLE_NOT_EQUAL(actual, expected)                             \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_double)
- #define CC_ASSERT_LONG_NOT_EQUAL(actual, expected)                            \
+ #define CC_ASSERT_LONG_NOT_EQUAL(actual, expected)                               \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_long)
- #define CC_ASSERT_LONG_LONG_NOT_EQUAL(actual, expected)                       \
+ #define CC_ASSERT_LONG_LONG_NOT_EQUAL(actual, expected)                          \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_long_long)
- #define CC_ASSERT_INT_NOT_EQUAL(actual, expected)                             \
-     __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_int)
- #define CC_ASSERT_INT8_NOT_EQUAL(actual, expected)                            \
+ #define CC_ASSERT_INT_NOT_EQUAL(actual, expected)                                \
+     __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((int)(actual), (int)(expected), test_type_int)
+ #define CC_ASSERT_INT8_NOT_EQUAL(actual, expected)                               \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_int8)
- #define CC_ASSERT_INT16_NOT_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_INT16_NOT_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_int16)
- #define CC_ASSERT_INT32_NOT_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_INT32_NOT_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_int32)
- #define CC_ASSERT_INT64_NOT_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_INT64_NOT_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_int64)
- #define CC_ASSERT_UINT_NOT_EQUAL(actual, expected)                            \
+ #define CC_ASSERT_UINT_NOT_EQUAL(actual, expected)                               \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_uint)
- #define CC_ASSERT_UINT8_NOT_EQUAL(actual, expected)                           \
+ #define CC_ASSERT_UINT8_NOT_EQUAL(actual, expected)                              \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_uint8)
- #define CC_ASSERT_UINT16_NOT_EQUAL(actual, expected)                          \
+ #define CC_ASSERT_UINT16_NOT_EQUAL(actual, expected)                             \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_uint16)
- #define CC_ASSERT_UINT32_NOT_EQUAL(actual, expected)                          \
+ #define CC_ASSERT_UINT32_NOT_EQUAL(actual, expected)                             \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_uint32)
- #define CC_ASSERT_UINT64_NOT_EQUAL(actual, expected)                          \
+ #define CC_ASSERT_UINT64_NOT_EQUAL(actual, expected)                             \
      __CC_ASSERT_NUMBER_NOT_EQUAL_TYPE((actual), (expected), test_type_uint64)
  
  /**
@@ -305,18 +305,18 @@
   * CC_ASSERT_STRING_NOT_EQUAL takes 2 strings and reports on their inequality. 
   */
  #define CC_ASSERT_STRING_NOT_EQUAL(actual, expected) \
-     do {                                                 \
-         if (strcmp(actual, expected) == 0) {             \
-             cc_result_t ccrt = (cc_result_t) {           \
-                 .filename = __FILE__,                    \
-                 .function = (char*)__FUNCTION__,         \
-                 .type = test_type_string,                \
-                 .result = false,                         \
-                 .line = __LINE__                         \
-             };                                           \
-             __CC_STRING_VAL_COPY(actual, expected);      \
-             return ccrt;                                 \
-         }                                                \
+     do {                                             \
+         if (strcmp(actual, expected) == 0) {         \
+             cc_result_t ccrt = (cc_result_t) {       \
+                 .filename = __FILE__,                \
+                 .function = (char*)__FUNCTION__,     \
+                 .type = test_type_string,            \
+                 .result = false,                     \
+                 .line = __LINE__                     \
+             };                                       \
+             __CC_STRING_VAL_COPY(actual, expected);  \
+             return ccrt;                             \
+         }                                            \
      } while (0)
  
  
@@ -333,23 +333,23 @@
  /**
   * CC_ASSERT_TRUE. 
   */
- #define CC_ASSERT_TRUE(actual)                               \
-     do {                                                     \
-         if (actual == false) {                               \
-             __CC_ASSIGN_BOOL(actual);                        \
-             return ccrt;                                     \
-         }                                                    \
+ #define CC_ASSERT_TRUE(actual)        \
+     do {                              \
+         if (actual == false) {        \
+             __CC_ASSIGN_BOOL(actual); \
+             return ccrt;              \
+         }                             \
      } while (0)
  
  /**
   * CC_ASSERT_FALSE. 
   */
- #define CC_ASSERT_FALSE(actual)                          \
- do {                                                     \
-     if (actual != false) {                               \
-         __CC_ASSIGN_BOOL(actual);                        \
-         return ccrt;                                     \
-     }                                                    \
+ #define CC_ASSERT_FALSE(actual)   \
+ do {                              \
+     if (actual != false) {        \
+         __CC_ASSIGN_BOOL(actual); \
+         return ccrt;              \
+     }                             \
  } while (0)
  
  /**
